config.h: increased exclude list to 128
On the Mac Mini 2012, there are 63 sensors and over 20+ of them are
unusable as part of the avg_temp calculation. Some of the are not
present, thus giving a read out of -127 C.

NOTE: this ia quick fix. The true fix should be to look at the actual
data read back from the hwmon and identifying the temperatures that are
invalid or with an error code (? if that's in the code).

But having the exclusion works is a good idea anyway. We have memory to
do more than 20 sensors on most modern Macs.