*HACK* config.c: fixed ill-formed line warning
This is not a real fix. Gets the problem around.  Config reads should
know the difference between a line a part of a line. Forcing a line to
end at 63 or 256 is not the way to go.

For now, this works, added warning to user about long lines > 256 so
that they'll have a clue as to what happened when their config settings
gets ignored.